+++ rg_instructor_analytics/views/__init__.py
@@ -1,3 +1,5 @@
 """
 Package with views.
 """
+
+# PEP8 violation - see: https://www.python.org/dev/peps/pep-0008/?#package-and-module-names
